Make tracking easier to configure and introduce tracking interceptors

+++ common/src/main/java/io/fluxcapacitor/common/api/tracking/Read.java
@@ -24,5 +24,5 @@ public class Read extends Request {
     String processor;
     int channel;
     int maxSize;
-    int maxTimeout;
+    long maxTimeout;
 }
